+++ src/conponents/Profile/Profile.js
@@ -1,8 +1,12 @@
 import React from 'react';
+import Header from '../Header/Header';
 
 function Profile() {
   return (
-    <div>Profile Page</div>
+    <div>
+      <Header />
+      <p>Profile Page</p>
+    </div>
   );
 }
 export default Profile;
