+++ src/conponents/SubmitForm/SubmitForm.css
@@ -1,6 +1,8 @@
 .form {
   margin-top: 100px;
   margin-bottom: 50px;
+  display: flex;
+  justify-content: center;
 }
 
 .input {
